+++ lib/crystal/view_helpers.rb
@@ -6,7 +6,7 @@ module Crystal
 
     def meta_tags(pattern = //)
       meta.tags_by_pattern(pattern).map{|t|
-        tag(:meta, :property => t.name, :value => t.value_for_context(self))
+        tag(:meta, t.name_key => t.name, :content => t.value_for_context(self))
       }.join("\n").html_safe
     end
   end