+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+### 0.9.4 2013-02-28
+
+* Fix bugs with displaying tags: a tag key name should be `property` for OpenGraph & `name` for anything else, value should be called `content`
+* Remove default value for `twitter:url` — it has an OpenGraph fallback to `og:url`
+
 ### 0.9.3 2013-02-03
 
 * Fix interpolation bug